pyproject.toml
[project]
name = "aigrc"
version = "0.1.0"
description = "AI Governance, Risk, and Compliance toolkit for Python"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Pango Labs", email = "engineering@pangolabs.ai" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
keywords = ["ai", "governance", "compliance", "risk", "ml", "langchain", "llm"]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "typer>=0.9,<1.0",
    "rich>=13.0,<14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.5",
    "ruff>=0.1",
    "black>=23.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=1.24",
]

[project.scripts]
aigrc = "aigrc.cli.main:app"

[project.urls]
Homepage = "https://github.com/aigrc/aigrc-python"
Documentation = "https://aigrc.readthedocs.io"
Repository = "https://github.com/aigrc/aigrc-python"
Changelog = "https://github.com/aigrc/aigrc-python/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aigrc"]

[tool.mypy]
strict = true
python_version = "3.10"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
